All product info pulling to product card
