+++ client/src/app.css
@@ -42,7 +42,7 @@
 }
 
 .related_sale{
-  font: red;
+  color: red;
   text-decoration: line-through;
 
 }
